#972: move TenantClientCustomization classes/tests into customization package
